requirements.txt
streamlit
moviepy
openai-whisper
google-generativeai
torch
ffmpeg-python
